eatgo-api/src/main/java/kr/co/fastcampus/springrestaurant/application/RestaurantService.java
package kr.co.fastcampus.springrestaurant.application;

import kr.co.fastcampus.springrestaurant.domain.MenuItem;
import kr.co.fastcampus.springrestaurant.domain.MenuItemRepository;
import kr.co.fastcampus.springrestaurant.domain.Restaurant;
import kr.co.fastcampus.springrestaurant.domain.RestaurantRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class RestaurantService {
    private final RestaurantRepository restaurantRepository;
    private final MenuItemRepository menuItemRepository;

    public RestaurantService(RestaurantRepository restaurantRepository, MenuItemRepository menuItemRepository) {
        this.restaurantRepository = restaurantRepository;
        this.menuItemRepository = menuItemRepository;
    }

    public List<Restaurant> getRestaurants() {
        return restaurantRepository.findAll();
    }

    public Restaurant getRestaurant(Long id) {
        Restaurant restaurant = restaurantRepository.findById(id);
        List<MenuItem> menuItems = menuItemRepository.findAllByRestaurantById(id);
        restaurant.setMenuItems(menuItems);
        return restaurant;
    }
}
